+++ Sumdays/app/src/main/java/com/example/sumdays/network/WeekAnalysisResponse.kt
@@ -52,7 +52,7 @@ data class NetworkEmotionAnalysis(
     val dominantEmoji: String,
 
     @SerializedName("emotion_score")
-    val emotionScore: Double,
+    val emotionScore: Float,
 
     @SerializedName("trend")
     val trend: String? // "decreasing"
